Update submission.txt
During the execution of the task, unfortunately, the node disconnected due to the fact that the server ran out of memory. I discovered this too late, so the running time is less than 720. Please accept the task, because my only mistake is that I did not check the performance of the node at the time :)